+++ src/app/player_input.rs
@@ -1,3 +1,4 @@
+// TODO: Replace this with https://github.com/nicopap/ui-navigation
 use bevy::prelude::*;
 use leafwing_input_manager::orientation::Direction;
 use leafwing_input_manager::prelude::*;
